requirements.txt
opencv-python==4.1.1.26
opencv-contrib-python==4.1.2.30
numpy==1.17.2
Pillow==6.1.0
pylint==2.3.1
Flask==1.1.1
mysqlclient==1.4.6
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
Flask-Cors==3.0.8